backend/pyproject.toml
[tool.poetry]
name = "famiglia-recipes"
version = "2.0.0"
description = ""
authors = ["Crossoufire <vince.delmas.38@gmail.com>"]
readme = ""
license = "MIT"
package-mode = false


[tool.poetry.dependencies]
python = "^3.10 || ^3.11 || ^3.12"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
flask-cors = "^5.0.0"
flask-mail = "^0.10.0"
flask-httpauth = "^4.8.0"
flask-migrate = "^4.0.7"
pillow = "^11.0.0"
pyjwt = "^2.10.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
